fix: clean up / refactor demo app (#1301)
