+++ examples/ui-demo/src/components/configuration/CornerRadiusOptions.tsx
