+++ examples/ui-demo/src/components/configuration/IllustrationStyleOptions.tsx
