+++ examples/ui-demo/src/components/eoa-post-login/EOAPostLogin.tsx
@@ -5,8 +5,8 @@ import {
 
 export function EOAPostLogin() {
   return (
-    <div className="flex flex-col">
-      <div className="border-border border radius-2 px-6 lg:px-10 py-11 max-w-[486px] w-full bg-bg-surface-default">
+    <div className="px-10 py-11 flex flex-col border-none lg:border-solid border border-border rounded-lg overflow-hidden overflow-y-auto scrollbar-none">
+      <div className="max-w-[486px] w-full bg-bg-surface-default">
         <EOAPostLoginContents />
         <div className="hidden lg:block">
           <EOAPostLoginActions />
