+++ examples/ui-demo/src/components/mint-card/NFT.tsx
