+++ examples/ui-demo/src/components/mint-card/ValueProps.tsx
@@ -1,9 +1,9 @@
 import { MintStatus } from "./MintCard";
-import { LoadingIcon } from "../../icons/loading";
-import { CheckIcon } from "../../icons/check";
-import { GasIcon } from "../../icons/gas";
-import { DrawIcon } from "../../icons/draw";
-import { ReceiptIcon } from "../../icons/receipt";
+import { LoadingIcon } from "../icons/loading";
+import { CheckIcon } from "../icons/check";
+import { GasIcon } from "../icons/gas";
+import { DrawIcon } from "../icons/draw";
+import { ReceiptIcon } from "../icons/receipt";
 
 type Props = {
   status: MintStatus;
