+++ examples/ui-demo/src/components/preview/MobileSplashPage.tsx
@@ -14,7 +14,7 @@ export function MobileSplashPage() {
   }, [breakpoint, isAuthenticating, openAuthModal]);
 
   return (
-    <div className="flex flex-col flex-1 h-auto max-h-[calc(100svh-100px)] box-content px-4 pt-0 justify-between">
+    <div className="flex lg:hidden flex-col flex-1 h-auto max-h-[calc(100svh-100px)] box-content px-4 pt-0 justify-between h-full">
       {/* Header Text */}
       <div>
         <h3 className="min-[320px]:text-3xl min-[390px]:text-5xl min-[640px]:text-6xl min-[640px]:mt-12 text-center font-semibold tracking-tight mb-4 leading-none">
