+++ examples/ui-demo/src/components/topnav/TopNav.tsx
@@ -17,7 +17,7 @@ export function TopNav() {
   }, []);
 
   return (
-    <div className="fixed lg:sticky left-0 top-0 right-0 flex justify-center py-5 bg-demo-bg lg:bg-demo-nav-bg shadow-sm z-10">
+    <div className="fixed lg:sticky left-0 top-0 right-0 flex justify-center py-5 bg-demo-bg lg:bg-demo-nav-bg shadow-sm z-20">
       <div className="flex items-center justify-between h-full w-full max-w-[1440px] mx-6 md:mx-7 lg:mx-10">
         <div className="flex justify-center items-center gap-2">
           {/* This link is used to refresh the page when the user clicks the logo */}
