+++ examples/ui-demo/src/components/user-connection-avatar/DeploymentStatusIndicator.tsx
