+++ examples/ui-demo/src/components/user-connection-avatar/UserAvatar.tsx
